fixed uint length in hmac_append_long_key and hmac_append_long_text functions

+++ src/hmac.c
@@ -146,7 +146,7 @@ void hmac_append_int_text(hmac_ctx_t *ctx, uint32_t value) {
 }
 
 
-/**                         hmac_append_int_key(hmac_ctx_t*, uint32_t);
+/**                         hmac_append_long_key(hmac_ctx_t*, uint64_t);
  *
  *  Requires:               - hmac_ctx_init(hmac_ctx_t*, uint64_t, uint64_t);
  *
@@ -159,12 +159,12 @@ void hmac_append_int_text(hmac_ctx_t *ctx, uint32_t value) {
  *                          HMAC.
  *  @param value:           64 bit integer value that needs to be appended in key chunks.
  */
-void hmac_append_long_key(hmac_ctx_t *ctx, uint32_t value) {
+void hmac_append_long_key(hmac_ctx_t *ctx, uint64_t value) {
     sha1_append_long(&ctx->sha1_ctx_key, value);
 }
 
 
-/**                         hmac_append_int_text(hmac_ctx_t*, uint32_t);
+/**                         hmac_append_long_text(hmac_ctx_t*, uint64_t);
  *
  *  Requires:               - hmac_ctx_init(hmac_ctx_t*, uint64_t, uint64_t);
  *
@@ -177,7 +177,7 @@ void hmac_append_long_key(hmac_ctx_t *ctx, uint32_t value) {
  *                          HMAC.
  *  @param value:           64 bit integer value that needs to be appended in text chunks.
  */
-void hmac_append_long_text(hmac_ctx_t *ctx, uint32_t value) {
+void hmac_append_long_text(hmac_ctx_t *ctx, uint64_t value) {
     sha1_append_long(&ctx->sha1_ctx_text, value);
 }
